fix: extract utils to test-utils.js, add edge case tests, remove Jest
